refactor: 매퍼 파일 리네임 및 유틸리티 함수 분리
