+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/dto/response/AfternoteDetailResponse.kt
@@ -6,7 +6,6 @@ import com.afternote.feature.afternote.data.dto.AfternotePlaylist
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.JsonNames
 
 /**
  * Server response for GET /afternotes/{afternoteId}.
@@ -18,10 +17,8 @@ data class AfternoteDetailResponse(
     @SerialName("afternoteId") val afternoteId: Long,
     @SerialName("category") val category: String,
     @SerialName("title") val title: String,
-    @SerialName("createdAt")
-    @JsonNames("created_at", "createdAt") val createdAt: String = "",
-    @SerialName("updatedAt")
-    @JsonNames("updated_at", "updatedAt") val updatedAt: String = "",
+    @SerialName("createdAt") val createdAt: String = "",
+    @SerialName("updatedAt") val updatedAt: String = "",
     @SerialName("credentials") val credentials: AfternoteCredentials? = null,
     @SerialName("receivers") val receivers: List<AfternoteDetailReceiver>? = null,
     @SerialName("processMethod") val processMethod: String? = null,
