+++ feature/afternote/data/src/main/kotlin/com/afternote/feature/afternote/data/mapper/AfternoteListItemDtoMapper.kt
@@ -1,0 +1,18 @@
+package com.afternote.feature.afternote.data.mapper
+
+import com.afternote.feature.afternote.data.dto.AfternoteListItem
+import com.afternote.feature.afternote.domain.model.Item
+
+/**
+ * Maps server DTOs to domain models at the boundary only.
+ */
+
+fun AfternoteListItem.toDomain() =
+    Item(
+        id = afternoteId.toString(),
+        serviceName = title,
+        date = formatDateFromServer(createdAt),
+        type = categoryToServiceType(category),
+    )
+
+fun List<AfternoteListItem>.toDomainList() = map { it.toDomain() }
